+++ Editor/Scripts/Extensions/EditorWindowExtensions.cs
@@ -5,10 +5,19 @@ using UnityEngine;
 namespace Unity.FilmInternalUtilities.Editor {
 
 internal static class EditorWindowExtensions {
-    private static Vector2 GetWindowSize(this EditorWindow editorWindow) {
+    internal static Vector2 GetWindowSize(this EditorWindow editorWindow) {
         Rect pos = editorWindow.position;
         return new Vector2(pos.width, pos.height);
-    }}
+    }
+    
+    internal static void Resize(this EditorWindow window, Vector2Int windowSize) {
+        Rect pos = window.position;
+        pos.width       = windowSize.x;
+        pos.height      = windowSize.y;
+        window.position = pos;
+    }
+    
+}
 
 } //end namespace