+++ Runtime/Scripts/Utilities/TimelineUtility.cs
@@ -105,6 +105,20 @@ internal static class TimelineUtility {
         outAsset = null;
     }
     
+//--------------------------------------------------------------------------------------------------------------------------------------------------------------
+    
+    internal static void DeleteInvalidMarkers<MarkerType>(TrackAsset track) where MarkerType : Marker, ICanRefresh {
+        List<Marker> markersToDelete = new List<Marker>();
+        foreach (IMarker m in track.GetMarkers()) {
+            MarkerType marker = m as MarkerType;
+            if (null == marker)
+                continue;
+
+            if (!marker.Refresh()) markersToDelete.Add(marker);
+        }
+
+        foreach (Marker marker in markersToDelete) track.DeleteMarker(marker);
+    }
 
 }